requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0
scikit-learn>=1.0.0
librosa>=0.8.1
pydub>=0.25.1
tqdm>=4.62.0
ipykernel>=6.0.0
jupyter>=1.0.0
soundfile>=0.10.3
xgboost>=1.5.0
plotly>=5.3.0
PyYAML>=6.0
pytest>=6.2.5
flake8>=4.0.0